# examples/t1api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UsageAggregateResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "group_by": {
            "title": "Group By",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/UsageAggregateRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "since": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Since"
          },
          "until": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until"
          }
        },
        "required": [
          "group_by",
          "rows",
          "count",
          "request_id"
        ],
        "title": "UsageAggregateResponse",
        "type": "object"
      },
...
      "UsageRemainingResponse": {
        "properties": {
          "input_remaining": {
            "title": "Input Remaining",
            "type": "integer"
          },
          "is_exhausted": {
            "title": "Is Exhausted",
            "type": "boolean"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "output_remaining": {
            "title": "Output Remaining",
            "type": "integer"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "model_id",
          "input_remaining",
          "output_remaining",
          "is_exhausted",
          "request_id"
        ],
        "title": "UsageRemainingResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Controlled gateway into HyperNix-pip. The client requests an operation; the server decides what exists, what's available, and how much is left. See wiki/T1-API.md for the full contract.",
    "title": "HyperNix T1 API",
    "version": "0.71.5b3"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/usage/by": {
      "get": {
        "description": "\"Usage by model / key / server / module / user / account\", as one\nendpoint rather than six near-identical ones.\n\n``group_by`` is checked against the storage layer's allowlist before\nit reaches SQL \u2014 it becomes an identifier, not a bound parameter, so\nit can never be free-form caller input.",
        "operationId": "usage_by_usage_by_get",
        "parameters": [
          {
            "description": "One of: model_id, key_id, server_id, module_id, user_id, account_id, endpoint",
            "in": "query",
            "name": "group_by",
            "required": true,
            "schema": {
              "description": "One of: model_id, key_id, server_id, module_id, user_id, account_id, endpoint",
              "title": "Group By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "range",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Range"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "in": "query",
            "name": "key_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Key Id"
            }
          },
          {
            "in": "query",
            "name": "model_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          },
          {
            "in": "query",
            "name": "server_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Server Id"
            }
          },
          {
            "in": "query",
            "name": "module_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Module Id"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageAggregateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Usage By",
        "tags": [
          "usage"
        ]
      }
    },
    "/usage/cost": {
      "get": {
        "description": "Actual cost with optional forecast \u2014 totals, date ranges, and\nper-model/per-server breakdowns in one shape.",
        "operationId": "usage_cost_usage_cost_get",
        "parameters": [
          {
            "description": "One of: model_id, key_id, server_id, module_id, user_id, account_id, endpoint",
            "in": "query",
            "name": "group_by",
            "required": false,
            "schema": {
              "default": "model_id",
              "description": "One of: model_id, key_id, server_id, module_id, user_id, account_id, endpoint",
              "title": "Group By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "range",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Range"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "in": "query",
            "name": "key_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Key Id"
            }
          },
          {
            "in": "query",
            "name": "model_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          },
          {
            "in": "query",
            "name": "server_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Server Id"
            }
          },
          {
            "in": "query",
            "name": "module_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Module Id"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "description": "Include a spend forecast for the horizon.",
            "in": "query",
            "name": "forecast",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include a spend forecast for the horizon.",
              "title": "Forecast",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "forecast_horizon_seconds",
            "required": false,
            "schema": {
              "default": 2592000.0,
              "exclusiveMinimum": 0,
              "title": "Forecast Horizon Seconds",
              "type": "number"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageCostResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Usage Cost",
        "tags": [
          "usage"
        ]
      }
    },
    "/usage/current": {
      "get": {
        "operationId": "usage_current_usage_current_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageCurrentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Usage Current",
        "tags": [
          "usage"
        ]
      }
    },
    "/usage/estimate": {
      "post": {
        "description": "What an operation *would* cost. Records nothing and reserves\nnothing \u2014 see ``t1api.cost``.\n\nPOST rather than GET because it takes a body and, more importantly,\nbecause a cost estimate for a specific prompt size is a question\nabout a proposed operation, not a resource that can be bookmarked.",
        "operationId": "usage_estimate_usage_estimate_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsageEstimateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageEstimateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Usage Estimate",
        "tags": [
          "usage"
        ]
      }
    },
    "/usage/history": {
      "get": {
        "operationId": "usage_history_usage_history_get",
        "parameters": [
          {
            "description": "1h|24h|7d|30d|all",
            "in": "query",
            "name": "range",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "1h|24h|7d|30d|all",
              "title": "Range"
            }
          },
          {
            "description": "POSIX timestamp, inclusive.",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "POSIX timestamp, inclusive.",
              "title": "Since"
            }
          },
          {
            "description": "POSIX timestamp, exclusive.",
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "POSIX timestamp, exclusive.",
              "title": "Until"
            }
          },
          {
            "description": "Admin only; ignored otherwise.",
            "in": "query",
            "name": "key_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Admin only; ignored otherwise.",
              "title": "Key Id"
            }
          },
          {
            "in": "query",
            "name": "model_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          },
          {
            "in": "query",
            "name": "server_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Server Id"
            }
          },
          {
            "in": "query",
            "name": "module_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Module Id"
            }
          },
          {
            "description": "Admin only; ignored otherwise.",
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Admin only; ignored otherwise.",
              "title": "User Id"
            }
          },
          {
            "description": "Admin only; ignored otherwise.",
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Admin only; ignored otherwise.",
              "title": "Account Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Usage History",
        "tags": [
          "usage"
        ]
      }
    },
    "/usage/remaining": {
      "get": {
        "operationId": "usage_remaining_usage_remaining_get",
        "parameters": [
          {
            "description": "Registered model_id to check remaining allowance for.",
            "in": "query",
            "name": "model_id",
            "required": true,
            "schema": {
              "description": "Registered model_id to check remaining allowance for.",
              "title": "Model Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageRemainingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Usage Remaining",
        "tags": [
          "usage"
        ]
      }
    }
  }
}